// frontend/src/api/generated/model/newsSourceRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ZeroAI
 * OpenAPI spec version: 0.1.0
 */

export interface NewsSourceRead {
  /**
     * @minLength 1
     * @maxLength 80
     */
  name: string;
  /**
     * @minLength 8
     * @maxLength 500
     */
  url_template: string;
  enabled?: boolean;
  id: number;
}