// frontend/src/api/generated/model/lLMConfigRead.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ZeroAI
 * OpenAPI spec version: 0.1.0
 */
import type { LLMConfigReadProvider } from './lLMConfigReadProvider';
import type { LLMConfigReadThinkingEffort } from './lLMConfigReadThinkingEffort';

export interface LLMConfigRead {
  provider: LLMConfigReadProvider;
  model: string;
  base_url: string | null;
  thinking: boolean;
  thinking_effort: LLMConfigReadThinkingEffort;
  api_key_set: boolean;
}